Changed RelationalAggregateRepository.kt to be flow first and changed data colums to Blob

+++ core/src/main/kotlin/com/rmpader/eventsourcing/repository/AggregateRepository.kt
@@ -36,7 +36,7 @@ interface AggregateRepository<E, S> {
 
     suspend fun deleteFromOutbox(eventIds: Set<String>)
 
-    suspend fun pollOutbox(limit: Int): Flow<OutboxRecord<E>>
+    suspend fun pollOutbox(limit: Long): Flow<OutboxRecord<E>>
 
     suspend fun cleanupStaleOutboxClaims(staleAfterMillis: Long): Long
 }
